step

+++ src/server/CrossDataBase.Server.Business.Abstraction/Core/Engine/IEngineService.cs
@@ -2,6 +2,6 @@
 namespace CrossDataBase.Server.Business.Abstraction.Core.Engine;
 public interface IEngineService
 {
-    Task RunAsync(long processId);
+    Task RunAsync(int processId);
     //Task StartAsync(long processId, long historyId);
 }
