+++ src/server/CrossDataBase.Server.DataAccess.Abstraction/Core/ProcessData/IProcessDataDbReader.cs
@@ -4,5 +4,5 @@ namespace CrossDataBase.Server.DataAccess.Abstraction.Core.ProcessData;
 
 public interface IProcessDataDbReader
 {
-    Task<ProcessDbModel> GetAsync(long id);
+    Task<ProcessDbModel> GetAsync(int id);
 }
